requirements.txt
# Core Dependencies - Hybrid PDF Extraction Pipeline
opendataloader-pdf>=1.0.0
google-genai>=0.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
supabase>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# File Handling & Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'  # Windows DLL for python-magic
httpx>=0.24.0

# Rate Limiting
slowapi>=0.1.9

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
